Fix parsing of KDC override in KerberosTransportBase (#53)
This change fixes a small issue when parsing the override string
given to KerberosClient for KDC name. The condition checking the
array length was incorrect (checking against 0 instead of 1). This
caused array IndexOutOfBoundException when user set the KDC string
without specifying a port number.

Additionally, the default port being used was 0 instead of 88 in
the case of KDC override. Added an else block to set the default.