Clean rented buffer during PAC generation.
Generate PAC is called twice before signing and after signing. Due to CryptoPool using ArrayPool<byte>.Shared internally it can retrieve an empty buffer or with some data. ( if previously rented buffer weren't cleaned on return) so the actual data in buffer may differ on each of the generations.

Because of this there is a variability in what data was signed and what will be returned as encoded binary pac.