Nullable handling and generate assembly info
Code written prior to C# 8 generally has warnings related to nullable
reference type warnings. Setting nullable as diabled for the projects
and just enabling it for the one place it is used in a more recent
update makes the warnings go away.

Generating assembly info allows the target for net8.0-windows to set
runtime versioning to Windows which avoids a bunch of CA1416 errors.